repoint requests page
