+++ web/pages/requests.tsx
@@ -1,30 +1,35 @@
+import { User } from "@supabase/auth-helpers-react";
 import { GetServerSidePropsContext } from "next";
 import AuthLayout from "../components/shared/layout/authLayout";
 import MetaData from "../components/shared/metaData";
-import RequestsPage from "../components/templates/requests/requestsPage";
+import RequestsPageV2 from "../components/templates/requestsV2/requestsPageV2";
 import { SupabaseServerWrapper } from "../lib/wrappers/supabase";
-import { useOrg } from "../components/shared/layout/organizationContext";
+import { SortLeafRequest } from "../services/lib/sorts/requests/sorts";
 
-interface RequestsProps {
-  user: any;
-  page: number;
+interface RequestsV2Props {
+  user: User;
+  currentPage: number;
   pageSize: number;
-  sortBy: string | null;
+  sort: SortLeafRequest;
 }
 
-const Requests = (props: RequestsProps) => {
-  const { user, page, pageSize, sortBy } = props;
+const RequestsV2 = (props: RequestsV2Props) => {
+  const { user, currentPage, pageSize, sort } = props;
 
   return (
-    <MetaData title="Requests">
+    <MetaData title={"Requests"}>
       <AuthLayout user={user}>
-        <RequestsPage page={page} pageSize={pageSize} sortBy={sortBy} />
+        <RequestsPageV2
+          currentPage={currentPage}
+          pageSize={pageSize}
+          sort={sort}
+        />
       </AuthLayout>
     </MetaData>
   );
 };
 
-export default Requests;
+export default RequestsV2;
 
 export const getServerSideProps = async (
   context: GetServerSidePropsContext
@@ -46,15 +51,19 @@ export const getServerSideProps = async (
   const { page, page_size, sort } = context.query;
 
   const currentPage = parseInt(page as string, 10) || 1;
-  const pageSize = parseInt(page_size as string, 10) || 25;
-  const sortBy = (sort as string) || null;
+  const pageSize = parseInt(page_size as string, 10) || 10;
+  const sortLeaf = sort
+    ? (JSON.parse(sort as string) as SortLeafRequest)
+    : {
+        created_at: "desc",
+      };
 
   return {
     props: {
       user: user,
-      page: currentPage,
-      pageSize: pageSize,
-      sortBy: sortBy,
+      currentPage,
+      pageSize,
+      sort: sortLeaf,
     },
   };
 };
